fill unit_settings
